avoid crash by dropping GAME_OVER event
Play Store reports say I'm getting this event when I don't have a
valid jni game instance to work with. So test for that, and drop the
event. If it's a race condition this should fix it. If not I'll get
some other crash report that might be more useful.