remove custom localization stuff
I had a scheme where, as on WinMobile, I could download localization
data based on the local of the device I was on. Saved a lot of
space. Google didn't care about space, and I haven't for a while
either, so it's coming out except for the API that did in-place
translation of UI elements.