Up MQTT client version, and min sdk to 19
I'm getting a lot of crashes deep in the MQTT client. Upgrading may not fix
it, but it does crash on launch for JellyBean and before. So this
release will have the new MQTT client AND not get installed on the six
or so users I have running too-old Android. If the crashes continue
perhaps I don't orphen them. TBD.