fixes to data export/import
Main problem, aside from the explanations being stubbed out, is that
zip file overwriting doesn't work: once an entry is in the file, it'll
be there when the file's overwritten even if that entry is
excluded. This happens even if I create the file in memory and then
overwrite using the bytes.