tweak nullability
There are a lot of jni calls that assert their params are non-null but
where the java api has nullable params. I need to fix that. This is
just the one case where an assertion is firing (though it makes no
sense and I don't think I'm fixing the problem....)