add Keep-Alive service to improve background receipt of MQTT
This adds a Service that tries to stay alive in the
Foreground (Android term) so that the MQTT connection doesn't get
killed by battery optimizations. The service seems to run forever IFF
the user has turned off all optimizations for the app. It puts up an
alert as required when running, and another when killed warning that
an app launch is required to check for moves. Still to do: turn it
off/on when events happen, e.g. when whether it should be running
based on how old the newest move is changes.